Provide basic implementation, support BC1/BC2/BC3
Provide a basic implemenation of the library with support for
BC1/BC2/BC3 (DXT1/DXT3/DXT5) texture decompression.

Decompression functions for other format are still skeletons.

Functions to decompress an entire texture in either tiled or linear
format have been implemented.

A test/validated program is provided, which comes with a set of sample
texture files.

Signed-off-by: Harm Hanemaaijer <fgenfb@yahoo.com>